+++ pages/index.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
-import { useFetchPokemonWithInfinityScroll } from './hooks/usePokemon';
+import { useFetchPokemonWithInfinityScroll } from '../hooks/usePokemon';
 import HomePage from '@/components/pages/HomePage';
 import MyBeatLoader from '@/components/UI/molecules/BeatLoader';
-import useFetchNextPage from './hooks/useFetchNextPage';
+import useFetchNextPage from '../hooks/useFetchNextPage';
 import Container from '@/components/UI/atoms/Container';
 
 const IndexPage: FC = (): JSX.Element => {
